requirements.txt
Flask==1.1.2
Flask-Caching
attrs==19.3.0
google-cloud-storage
jinja2
parameterized
protobuf==3.11.3
pytest
requests==2.23.0
six
wheel